Moved fidelity setter to super class.
